Temporary switch over to AOSP camera.
CM borked their fork of the camera application. It has a rather nasty crash when changing from video to still mode with the front camera. Most of their other changes to the camera application really don't apply to us to begin with, so nothing lost by going AOSP on this one, which isn't broken.