Back to CM camera
A problem with our camera blobs causes the video encoder library to be told the wrong resolution without a CM camera hack. This bug is worse than the one that was solved by switching to upstream camera, since it effectively meant that video recording did not work at all. It only pretended to.